feat: fix bugs

+++ src/modules/person/services/person.service.ts
@@ -1,5 +1,6 @@
 import { CreatePersonV2Dto } from '@/modules/person/dto/create-person-v2.dto';
 import { CreatePersonDto } from '@/modules/person/dto/create-person.dto';
+import { FindPersonPaginatedDto } from '@/modules/person/dto/find-person-paginated.dto';
 import { UpdatePersonV2Dto } from '@/modules/person/dto/update-person-v2.dto';
 import { UpdatePersonDto } from '@/modules/person/dto/update-person.dto';
 import { Person } from '@/modules/person/entities/person.entity';
@@ -14,46 +15,77 @@ export class PersonService {
     private readonly personRepository: Repository<Person>,
   ) {}
 
-  async create(createPersonDto: CreatePersonDto): Promise<Person> {
+  async create(createPersonDto: CreatePersonDto, userId: number): Promise<Person> {
     const duplicatedPersons = await this.personRepository.createQueryBuilder('p')
-      .orWhere('p.cpf = :cpf', { cpf: createPersonDto.cpf })
-      .orWhere('p.email = :email', { email: createPersonDto.email })
+      .andWhere('p.userId = :userId', { userId })
+      .andWhere('p.cpf = :cpf OR p.email = :email', { cpf: createPersonDto.cpf, email: createPersonDto.email })
       .getCount();
     
     if (duplicatedPersons > 0) {
       throw new NotAcceptableException(`Person with cpf or email already exists!`)
     } 
 
-    const person = this.personRepository.create(createPersonDto);
+    const person = this.personRepository.create({ ...createPersonDto, userId });
     return this.personRepository.save(person);
   }
 
-  async createV2(createPersonDto: CreatePersonV2Dto): Promise<Person> {
-    return await this.create(createPersonDto);
+  async createV2(createPersonDto: CreatePersonV2Dto, userId: number): Promise<Person> {
+    return await this.create(createPersonDto, userId);
   }
 
-  async findAll(): Promise<Person[]> {
-    return this.personRepository.find();
+  async findAll(findPersonPaginated: FindPersonPaginatedDto, userId: number, returnAddress?: boolean): Promise<{ persons: Person[], pagination: { itemsPerPage: number, page: number, totalItems: number, }}> {
+    const page = findPersonPaginated.page ?? 1;
+    const itemsPerPage = findPersonPaginated.itemsPerPage ?? 10;
+
+    const offset = (page - 1) * itemsPerPage;
+
+    const qb = this.personRepository.createQueryBuilder('p');
+
+    if (returnAddress) {
+      qb.leftJoinAndSelect('p.address', 'address');
+    }
+
+    if (findPersonPaginated.personName) {
+      qb.andWhere(`p.name ilike TRIM('%${findPersonPaginated.personName}%')`)
+        .andWhere('p.userId = :userId', { userId });
+    }
+
+    qb.orderBy('p.createdAt', 'DESC');
+
+    const totalItems = await qb.getCount();
+    const persons = await qb
+      .skip(offset)
+      .take(itemsPerPage)
+      .getMany();
+
+    return {
+      persons,
+      pagination: {
+        itemsPerPage: Number(itemsPerPage),
+        page: Number(page),
+        totalItems,
+      },
+    };
   }
 
-  async findOne(id: number): Promise<Person> {
-    const person = await this.personRepository.findOneBy({ id });
+  async findAllV2(findPersonPaginated: FindPersonPaginatedDto, userId: number) {
+    return await this.findAll(findPersonPaginated, userId, true);
+  }
+
+  async findOne(id: number, userId: number): Promise<Person> {
+    const person = await this.personRepository.findOneBy({ id, userId });
     if (!person) throw new NotFoundException(`Person with ID ${id} not found`);
     return person;
   }
 
-  async update(id: number, updatePersonDto: UpdatePersonDto): Promise<Person> {
-    const person = await this.personRepository.findOneBy({ id });
+  async update(id: number, updatePersonDto: UpdatePersonDto, userId: number): Promise<Person> {
+    const person = await this.personRepository.findOneBy({ id, userId });
     if (!person) throw new NotFoundException(`Person with ID ${id} not found`);
 
     if (updatePersonDto.cpf || updatePersonDto.email) {
-      const query = this.personRepository.createQueryBuilder('p');
-      if (updatePersonDto.cpf) {
-        query.orWhere('p.cpf = :cpf', { cpf: updatePersonDto.cpf });
-      }
-      if (updatePersonDto.email) {
-        query.orWhere('p.email  = :email', { email: updatePersonDto.email  });
-      }
+      const query = this.personRepository.createQueryBuilder('p')
+        .andWhere('p.userId = :userId AND p.id <> :id', { userId, id: person.id })
+        .andWhere('(p.cpf = :cpf OR p.email  = :email)', { cpf: updatePersonDto.cpf, email: updatePersonDto.email });
 
       const duplicatedPersons = await query.getCount();
       if (duplicatedPersons > 0) {
@@ -64,11 +96,14 @@ export class PersonService {
     return this.personRepository.save({ ...person, ...updatePersonDto });
   }
 
-  async updateV2(id: number, updatePersonDto: UpdatePersonV2Dto): Promise<Person> {
-    return await this.update(id, updatePersonDto);
+  async updateV2(id: number, updatePersonDto: UpdatePersonV2Dto, userId: number): Promise<Person> {
+    return await this.update(id, updatePersonDto, userId);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, userId: number): Promise<void> {
+    const person = await this.personRepository.findOneBy({ id, userId });
+    if (!person) throw new NotFoundException(`Person with ID ${id} not found`);
+
     const result = await this.personRepository.delete(id);
     if (result.affected === 0) throw new NotFoundException(`Person with ID ${id} not found`);
   }
